requirements.txt
Django==2.2.16
djangorestframework==3.13.1
djangorestframework_simplejwt==4.7.2
pytest==6.2.4
